Refactor interface generators
This is a significant refactor to make the module easier to understand
and work with. The various very specialized functions were split out
into more generic functions for easier understanding of the build graph,
and some aspects were split further/reused. In the end, the build graph
should look like:

IDL ----> TypeDescription
 |   `----|---------,  |
 |        v         v  v
 |--> CGenerator -> CppGenerator
 |        v
 `--> PythonGenerator

To match the changes here, providers in other files have been renamed
to remove "Aspect" from "AspectInfo", as it is redundant.