interfaces: Allow .idl files in ros2_interface_library
This so far seems to only be used by some rosidl tests, but it's good to
have it nevertheless. We currently classify all .idl files as messages;
to do it properly would require significantly more effort (we
essentially need to parse the IDL file and see how many interfaces it
contains, which is what the generators do).